Changes Hearing Model validation script
Changes the validation script 'hearing_model_validation' to clarify some of the calculations, and adds functions that calculate the time-dependent loudness 'N(l)' and the single-value loudness 'N' as per Eqs. 116 and 117 of the ECMA-418-2, 2nd Ed (2022). The standard specifies a power average (and not a common 'mean' operator) to obtain the single-value loudness, and a missing factor of 'delta_z = 0.5' is added in the integration of the specific loudness over all critical bands.